Tag system.
A new tagging system was added. When an important event happens,
it is annotated on the timeline. Later, when viewing the recording
a message appears on the screen explaining the event.
For now, it's used to indicate changes in SLAM/mapping modes.

To be explored: plot these tags along the errors, let the user
add interactively a new tag.

For now, the user may directly change the tags.out file, which is
pretty simple, though inconvenient.