User tags.
A new key binding has been added: using the G key, the user can
create a new tag at the current time. The message is just "User tag",
but it eases the pain of finding the exact times of important events.
More meaningful names can be created later by modifying the output
file itself.

Since the Viewer usually is not expected to change the files it works
with, since it is a *viewer*, when tags are added and the file is
rewritten, a backup of the old record is saved.