Kinect descriptor filtering.
To try and improve the quality of the measurements, a rough outlier
rejection scheme is implemented in the Kinect image processing
stage. It matches keypoints with the previous frame and uses RANSAC
to find the best candidate motion, removing outliers.
Since this scheme is slow, it can be turned off with a flag.
Similarly, the sidebar can now be disabled to speed the system as
much as possible (this removes the need for duplicating internal
frame arrays).

Some very old code in the KinectVehicle class was removed as it
has nothing to do it with the current approach.