Prerecorded Tags for SLAM/Mapping.
The RecordVehicle class now also uses prerecorded tags to figure out
when to change from SLAM to Mapping mode and vice versa. This is
important to remove any advantage of an algorithm over another,
since these seconds are key to a fixing the euclidean freedom of
the problem, which can make the final errors vary wildly.