ForceSLAM/Mapping before frame update.
The order between the SLAM update and the forceSLAM/Mapping calls
has been inverted so user input acts immediately. This is so the
system can trigger a mapping instruction at the very first frame,
which otherwise would be oddly off on the plots. No big changes
happens because of this (as the initial mapping stage in the
simulations was long enough to make the first frame irrelevant).
An alternative would be to call the program with the onlymapping
flag on, but that is less intuitive.