Map abstraction.
In preparation for giving k-d trees another shot, the Lists of
Gaussian have been replaced with an Map class, which will soon
contain the k-d tree routines.
GraphSLAM requires data association, so the list index is important
in that approach. Therefore, an additional IndexedMap that acts
just like a regular list is implement. Both classes are related
by the IMap interface.