Moved headless timer outside.
When running a headless simulation, a timer is introduced to see
how long the entire simulation takes. This was performed by the
Simulation class. Now it only does the simulation and the time
is measured by the calling function, in the Program class.
Headless can help in other areas (smoothing) so writing the time
every run can be inconvenient.