NoTerminate and simulation constructor.
A new boolean has been introduced on the Simulation constructor
which makes the simulation continue forever, even if the command
list has depleted or the vehicle has signaled it wants to stop.

To simplify internal construction of the Simulation class, the
navigator object must be supplied as a parameter and won't be
initialized inside the class. This will help when more complex
navigators require longer initialization. Note that it is the user's
responsibility to check that the navigator and the vehicle make
sense together.