Removed Q2LieJacobian singularity.
If the quaternion was unity, one of the rows of the jacobian became
negative infinity. Although mathematically correct, this can make
algorithms diverge unnecessarily. This case has been changed to
output zero, so the difficult coordinate is only ignored instead
of NaNing everything on its way. Since most of the time only the
six-dimensional covariance is important this does not affect the
performance of the system.