New keypoint extraction.
A new keypoint extraction library project has been added which uses
ORB features instead of FAST. It uses OpenCV to extract them. Note that
both methodologies use the same points, but ORB can order them (and
select the best n ones) in a better way due to the Harris scoring
and a pyramid scheme, albeit it is slower.