Postanalysis to C#.
Python is too painfully slow for OSPA. The results are identical
so no drawbacks. The final plotting should still be done in python.
Alongside location error and OSPA errors, new rotation and odometry
errors have been added. Also, the map size has been compared with
the real size and the location error is compared with the real
arclength of the path (to give a global error reference).
To compare the map sizes, the main program also outputs the visible
map at every time step.