No odometry map output.
Odometry now does not output its map estimate. Since it is so big,
it bogs down the rest of the process and it is not really that useful
(other than for visualization).
The headless simulation loop now outputs messages each time a
time step is completed. This way, the user may extrapolate the
time it will take to complete.

Some configuration values have been changed.
Some debug code in Pose3D has been removed.