Modified correct visible map size.
Previously, for calculating the visible map (to compare its size
with the output of the SLAM algorithm), all landmarks that were
inside the visible frustum were considered.
This may not be fair, since undetected landmarks can't really be
considered by the algorithms. A better comparison would not consider
such visible-but-not-detected landmarks, which has been implemented.