Script files into project.
Plotting script files have been moved into the main project folder.
They are now synced with the project using the solution file.
The result is the same, but it is easier to track everything from
the same editor (otherwise it gets confusing to remember where
changes were made, in the bin folder or the original file), and
it copies the file even if there's nothing else to do for the build
(previously copying may have required a complete rebuild to force
the copying script to run).

Note that config files are no longer tracked (they change too often).

Also, mapping simfile has been changed since it is run for longer.
It contains more landmarks than the rest.