Bugfix: double deletion in unmanaged isam2.
An update to MonoGame changed the internal behavior of the Run method
which now disposes the game window (which disposes its children).
Since the window was already being disposed by a using() statement,
the Dispose method would be run twice. Eventually the unmanaged
isam2 navigator object would be deleted twice.
Since the program should not depend in the internals of MonoGame,
the using() should be maintained.
Instead, the ISAM2Navigator Dispose method checks for null before
deleting the unmanaged object.