Bugfix: isam2 odometry.
iSAM2 and this project use a slightly different definition of
odometry, which means that passing odometry directly to the iSAM2
solver would not follow the track correctly (the error was small,
but nonzero). Now the estimated new pose is passed instead, so the
solver can calculate the odometry in its own terms.

This also exposed a bug in the Quaternion.Sqrt() method, which could
become undefined for a 360º rotation (q = -1). This rotation could
occur if quaternion were converted from matrices, since in such
conversion the sign is arbitrary.