Tuned mapping error plot.
The post-processor now also outputs the spatial component of the
map OSPA error. This way, it is easy to see how much of the error
is due to cardinality and how much is due to spatial distances.

Also, the map estimate was reverted back to filtering components
with weight higher than 0.8, which although inconsistent, seems to
work better.

The plotter python script has been corrected to allow more lines
in the same plot (defined styles for them) and a wording error was
removed.