Bugfix: video output.
Video output was suddenly missing (a long time ago). It probably was
due to some fixes introduced in the MonoGame framework (Graphics
object is now disposed correctly and sooner), which meant that at
the time of saving the avi file, the output would be pitch black.
Now the info is stored directly as Color[] instead of MonoGame's
Texture2D. This is better anyway, due to the structure of the code.

A bug in the side buffer size was corrected. The whole Size info
is now part of the Manipulator, instead of the vehicle (since it
is trivially calculated from the data).