Kinect tracker border.
While extracting features from the Kinect color images, there is
an area around the border which can't be sensed due to the size
of the patches required by the keypoint extraction. Hence, the
tracker must be adjusted to give such a border a zero probability
of detection, otherwise, all landmarks would eventually fade away.