Bugfix: Correct-step relevant map components.
The correct step in the bayesian update requires an effect on every
"landmark" for each measurement. Naturally, this can be approximated
and most of these relations may be dropped as the effect is close to
zero. So far, the selection of relevant components was done in the
measurement space, which isn't isotropic, so the chosen components
where not the best ones.
The system now uses the 3D vehicle space to choose these components,
obtaining more natural results.

Bugfix: a map Near() method could throw NullException due to the
structure of the internal KDTree implementation.