Subsample to Lowpass.
The Subsample step in the Kinect sensor image processing has been
replaced with a lowpass filter (averaging convolution), to avoid
suffering from aliasing.

When rendering the depth map, the system would change any zeros to
NaNs so the output mesh would be displayed correctly. This is no
longer necessary, since the new averaging takes care of it.

The default range for Kinect has been changed from 2m to 4m
(which corresponds with the real device).