Checkpointing.
Some simulations take really long so it is difficult to tune
parameters since it is necessary to stop the run first to see the
results, therefore losing the progress so far. Now the system
automatically saves the state to file every 15 minutes. This
way, the results can be observed and the simulation can be kept
running based on how good they look.