Map specific FindAll implementation.
IMap defines FindAll which returns an IMap object. Map implements
the interface, so defines such a method itself. Now, it also defines
a method of the same name which returns a Map (the concrete class)
instead. The generic implementation can still be accessed through
the interface. This can reduce unnecessary castings.