New FakeVehicle and FakeNavigator.
Previously, tampering with the vehicle position was done by changing
its public Pose property. This wasn't nice because the property could
just be a temporary calculation.
The difference is now clear-cut as mutable poses are only allowed in
the special subclass FakeVehicle (and FakeNavigator for the similar
case of estimated trajectories). Some minor changes have been introduced
in Viewer and Simulation to adapt to the new interface.