Bugfix: DefaultValue propagation and Murty's hang.
A couple of methods in SparseMatrix didn't respect the default value
of the matrix (Transpose() and Compact()), which now remains the same
after the operations. The same for the connected components function
in GraphCombinatorics

These errors and some ambiguity in the default value required for
the linear assignment made the Murty's enumerator to hang forever
retrieving the same pairing. Everything is correct now and for
the system to be quick, -Inf is assumed as the default value for
reduced profits in the Murty's algorithm. Although, it technically
could be anything, any other value becomes slow with big matrices
and it is what is used in the PHD anyway.

Added explicative note about why the last quadrant of PHD SetLikelihood
profit matrix is filled at a later stage (so connected components works
correctly).