Stability oriented Log-operators.
In preparation for other changes, two operators have been added
to the MatrixExtensions class.
LogSumExp calculates a logarithm of the sum of exponentials, which
usually appears when taking logarithms, since just like multiplications
become sums, sums become LogSumExps. A special function is required
since the change in orders of magnitude induced by the exp and log
make a naive implementation very numerically unstable.
Similarly, TemperedAverage calculates a "center of mass" for a
set of values, where each component is weighted proportionally to an
exponential weight. In this case, a naive implementation also suffers
from unstability.