Separated SetLikelihood to access log-likelihood.
The loopy system needs to calculate the log-likelihood log P(Z|X, M)
but if calculated naively it becomes numerically unstable.
Since most of the internal calculations are in the log space anyway,
the SetLikelihood function has been separated into a SetLogLikelihood
and the SetLikelihood functions (the latter just gets the exponential
of the former). In this arrangement, the system is much more stable.

An unnecessary temporary array has been removed from this processing,
the profit matrix calculation has also been moved into its own function
and the threshold Mahalanobis distance to be considered "too far"
has been raised from 3 to 5.