SetLogLikelihood visibility separation.
SetLogLikehood didn't consider visibility for each landmark; instead
it used the same maximal visibility for every one.
This is not correct according to the model (although the difference is
only relevant around the border of the visible field), so has been updated
to consider the particular visibility for each landmark.

A new method has been added which is pretty similar to SetLogLikelihood,
but with two big differences: It also computes the symbolical jacobian
of the whole function and it disregards visibility (assumes everything
is equally and maximally visible, the previous behaviour).
This method becomes useful for gradient descent, since the function
is stiff and numerical derivatives are not practical. Also, when
looking for interesting maxima, visibility is usually uninteresting
(beyond the basic filtering of which points are visible and which are
not) and swamp the whole thing, therefore making Loopy PHD impractical.