Bugfix: SetLikelihood misdetection and pairing incorrect values.
The misdetection quadrant of the profit matrix was filled with
log(PD) when it should be log(1-PD).
The pairing enumeration was incorrect for the lexicographical method.
Since it only reached a maximum of 200 it would stop before the 8!
combinations were used. The lexicographical method does not decrease
linearly. It may happen that the last combination had all the
likelihood and therefore the returned value may be meaningless, randomly.
The threshold for Murty's application has been lowered so 5! < 200,
and in this case there's no escape from enumerating all the combinations.