Lie-Odometry-Global array separation.
To simplify a lot of the math and readability of the code, the lie
algebra has been redefined in a simpler more textbook manner. This
way derivatives and other operations become much easier to reason
and concise to write. The old definition now has become the "odometry"
alternative, e.g. a pose has both Add and AddOdometry.
A third alternative (only for basic operations like Add and Subtract)
is the global space, where location and orientation are completely
independent (i.e. moving ahead is always north, not the bearing of
the vehicle).

Since now jacobians can be calculated in Lie space, there's no need
for MotionCovarianceQ, since local changes happen in 6 dimensions.
As a consequence, a lot of utility functions are no longer relevant.

The jacobian test has been completely changed in the Quaternion test
suite (it test better invariants and was possible to write thanks to
the separation above).

Bugfix: There were duplicate (and inconsistent!) definitions for
getting the rotation matrix from the quaternion. They have been
consolidated in the Quaternion class. One of the definitions
corresponded to the conjugate of the other, so the places where it
was used have been updated to use the conjugate (and it made sense
theoretically).