Updated external references.
Updated DotImaging library. The new version broke video reading,
due to an internal bug with a list of supported formats (at runtime,
the wrong one is selected and the library thinks it cannot read avi,
even though it can).
A new class works around this issue and manages to read the file
by forcing the correct list of supported formats.