Removed double frame grab in KinectVehicle.
Du to the particular way the TUM video was saved (unaligned depth
and color streams), the system would drop every other frame of
each stream, so as to align them.
Now, every frame is used and the ONI file must be properly aligned
to begin with. This way, both aligned and unaligned can be processed
by just preprocessing correctly. Although it would be desirable to
be able to distinguish both cases on the reader function, the
library does not expose the necessary details of the sensor file.

Note that it is not clear how a real sensor would interoperate
with this new behaviour (or the old one).