Plot error plot reference times.
 For absolute error plots, an optional reference time can be given
as a command line argument, where each pose is compared with the
pose at the the reference time. This way we may omit the effect of
errors previous to such reference time and get a feeling of how
certain events affect the performance of the algorithm.

For relative error plots, instead of using the pure odometry (i.e.
one frame difference), the plotter now calculates the difference
between ten frames as a way to reduce the noise.

Some internal methods now use poses instead of array for clarity.