Generic vehicles.
Major overhaul to the system by introducing generics into the
Vehicle and Navigators, so new state transition and sensor model
can be used instead of the basic Pose3D odometry and Pinhole Camera.

Several methods have been shifted around to make the interfaces and
abstract classes more general. The major one is the  newly introduced
interface IMeasurer, which contains everything related to both
Pose and Measurement (Vehicle now only contains a reference to this).

To add new models, a new IMeasurer needs to be created and corresponding
IPoses and IMeasurements. Also it needs to be registered in the
DynamicsModel enumeration, and the main program should be made aware of it.