Arbitrary map dimension.
The Map class can now use an arbitrary KDTree dimension, instead
of always choosing three. The change is trivial (and most of the time
the landmarks are in 3D), but it fixes one case in the PHD algorithm
where landmark are saved in a Map by their measured values, which
may have a different dimension. It was no problem before because
previous measurements had also three components, but in the 2D linear
model, they have two, invalidating the Map, which eventually throws
an exception.