Bugfix: Config file array reading.
The configuration class file reader was missing one type of field:
the double[], which is used to specify the VisibilityRamp, which
has been introduced relatively recently.
It was impossible to specify this parameter so it was stuck at the
default (and this was not ok for differently sized measurement spaces
as the new linear 2D measurements).