Model measurer into config file.
Instead of relying in a command line flag, the measurer must be
specified in the config file. This is to make the output file
transparent (otherwise it could be ambiguous which one was used).

Note that the internal config file of a record session is not opened
until later in the process so it can't be used to deduce the
measurer type. This could be fixed in the future by opening the zip
file first to inspect the configuration (or creating a non-global
configuration object, but that seems overkill). This also concerns
other global constants. For now it's important to re-specify the file,
like in all the included runnable scripts.

Added LoopyPHDNavigator tests for the Linear2D case.

Bugfix: LoopyPHDNavigator's render system was assuming a fixed
3D model for extracting covariances. This has been extended to any
model, as long as the first components correspond to x, [y, [z]],
where the last ones are optional. A similar fixed length was
corrected in RecordVehicle.