Robust missing file treatment.
Instead of just throwing an exception, the system prints an error
message when a file is not found.

Removed an unnecesary argument for the specialized runner function
in the main program.

The finalization routines (kinect and signal-interception shutdowns)
have been moved from the specialized runner function into the main
function.