Trajectory to Odometry into C#.
Maintaining two codebases for basically the same behaviour
(odometry <-> trajectory) in two different languages is not
sustainable. So, the python util script has been replaced with a
simpler C# function inside the main lib. This can be used with the
mono csharp REPL to more effectively manage the transformation.

Now, it is easier to reason about the results (since the transformation
inverse relationship is unit tested) and to use the full extension
of the library to explore them.

Note that the new function is less powerful (no depth alignment)
but that functionality is rarely used and can be emulated in a
few lines of interactive code.