Gradient ascent fixed gradient.
The gradient ascent used in the Loopy PHD system tries to move in
the gradient direction. If the new position is lower, it tries with
a smaller rate. Due to the structure of the code, every time the
set likelihood is calculated, the gradient is also calculated, and
it was overwriting the previous gradient. This meant that every time
the multiplier was reduced, the gradient could change, messing with
the effective rate (e.g. the gradient could grow at the same time the
multiplier shrunk, invalidating the update). It is better to
definitely separate both effects and only use the first gradient
(just and exponential search in the gradient direction).