Loopy multi-gaussian messages from map.
Messages from the map to the trajectory are too approximated to
be of real use (too many times they are overconfident). The resulting
algorithm works for some very mild noise regime, but oscillates heavily
or diverges if any map message is too wild.
Now the set likelihood is represented by a gaussian mixture which
is merged into one big gaussian (also considering the clutter
noise floor and the relative position of the past and future gaussians).

The new map-to-pose proposal is much more reasonable, although it still
needs some work.