Bugfix: PHD weight alpha missing an exponential.
The PHD weight re-weighting should multiply by P(Z|X, M) P(X|Z')/P(X|Z)
where Z' corresponds to the predicted distribution and Z to the
corrected distribution, both of which (using the PHD approximation)
are of the form prod(PHD(m) for m in map)/exp(mapsize).

The system was using prod(PHD(m) for m in map)/mapsize, missing the
exponential of the denominator, which has been added in the new version.

Also, the many multiplications and divisions have been moved into the
log-space for better accuracy.