Gaussian additional methods.
Three new methods have been added to the Gaussian class:

Bhattacharyya distance can be calculated between gaussians. This
can help merge components symmetrically (although it is much more
heavyweight than using the Mahalanobis distance).

Gaussians can be multiplied together, which is similar to fusing,
but the weight is updated appropriately to account for the reduction
in height of the result.

Gaussians can also be reweighted directly, instead of having to
construct a new one with the original mean and covariance.
This method does not change the immutable object. It generates a
new one but it does not re-compute the canonical representation from
scratch. Instead it just copies it from the original.

Bugfix: the main constructor and the Canonical factory constructor
were slightly inconsistent. The canonical version was using the
determinant while the main constructor was using the pseudo-determinant.
Now both of them use the pseudo-determinant, which helps in
under-determined systems.