Map DensityDistanceThreshold removed.
The DDT config value has been removed from the Map class. It doesn't
make sense as a general threshold since maps can be in different
spaces (e.g. measurement or position). Its use as a default is
error-prone, since it may not be a reasonable one. Now a specific
radius must be specified for every operation.

In the case of evaluation, not specifying it evaluates on the whole
components.

The global config value is still used, but in the Navigator class.