VisMaps propagation through recording.
When using a RecordVehicle, the visible map log was empty, because
the "real" methods weren't called, so no visibility was not calculated.
Since the original simulation had already calculated these VisMaps,
it is now read from the file and becomes accessible in the new
saved file.

Bugfix: In the Plot system, the map file reader was being used with
the incorrect landmark dimension.