Marginal multi-object estimator.
The MAP-like estimator for the PHD and LoopyPHD inferences has been
changed from the ad-hoc weight > 0.8 threshold to the Marginal
multi-object estimator, which calculates the expected size N and
extracts the N biggest components.
This had already been tried before, but seems to work great now that
a lot of bugs have been corrected.

In the OSPA metric calculation, some indices have been inverted for
consistency. The new matrix is equivalent and gives the same result.

Other minor style corrections.