Bugfix: visibility ramp simulation.
The simulated vehicle was not considering the visibility ramp when
measuring from the landmarks, which meant that there was an
inconsistency between the model and the estimation, which would
overestimate the weights of landmarks in around the visibility edge.

Now it detects landmarks proportionally to their probability of
detection, providing better results for the estimation process,
especially the expected map size.