Bugfix: Kinect vehicle crash.
The kinect implementation was a little behind and didn't run (crash
as soon as the program was started).

The specialized KinectMeasurer construction wasn't being used so
the system refused to start (this specialization uses the depth
map to inform the visibility model).

Also, when reading from the depth map to find out if a landmark was
visible, a missing pixel is now encoded as NaN, which would corrupt
all the calculations from there on, eventually crashing when the
particles couldn't be sorted by weight (NaNs are not comparable).
This encoding is used to make it easier to propagate missing pixels
when averaging the input. The visibility model now responds
not visible for NaN values.

These NaNs error weren't succesfully caught or thrown because the
Sort method failed with an InvalidOperation exception, which would
be caught as part of the gtsam exception handling. Since these case
was not expected (and appropriately flagged), an attempt to access
an inexistent detail entry would throw a null exception.
Now the appropriate null check has been introduced.

The Pose3D constructor now forces the orientation quaternion to
be normalized. Otherwise operations on it may be incorrect.
This should not be a problem inside the program, since every
construction is controlled and should already be correct, but the
user must specify the initial pose, which can then introduce
inconsistencies.