Screenshot mode.
A new user input has been enabled to take screenshots in high
resolution. Both the simulator and the viewer can trigger this
behavior. The auto-input can also specify it and change the
camera parameters to take it from the correct perspective.

The viewer also has a new screenshot mode which skips to certain
frames specified by tags and takes appropriate screenshots.

Note that, just as with exporting the sidebar video, it requires a
graphics context, so it is not available in headless mode.