Bugfix: possible map history corruption.
The map history was saved as a straight reference to the particular
map in time, which meant any subsequent modification to this map
would improperly modify the history.

Now it is duplicated, so it's not linked to the algorithm data in
any way, hence becoming immutable.

Unrelated: the rendering order has been reversed for the navigator,
because it gives a better look.