Online and screenshot visualization corrections.
The logic to compute if the algorithm was online was completely
wrong (and actually it usually produced the exact inverse answer).
It has been fixed, as well as the rendering logic for this case.

Similarly, when accounting for the screenshot mode of visualization,
the regular mode became incorrect. This includes a secondary path
for the regular full-recording visualization.