Removed unexplored delay.
Previously, the unexplored array passed to the predict step of
the PHD algorithm was delayed by one frame to help smooth out
false alarms, which can become spuriously activated because of this.

In the case of Loopy PHD, this doesn't really matter, because the
system looks at the entire picture, future and past, and it makes
the model less correct, so it is removed. The regular PHD still
works fine, but of course false alarms become slightly more common.